test(cfg): add bad_cfg_discovery case for no-target-spec-json
An additional rustc query is made in Cargo to collect the
target-spec-json from rustc to determine if the build target supports
compiling the standard library. As this functionality relies on a
nightly rustc, Cargo does not output any errors when parsing, and
continues as normal.

This commit adds a new test case to bad_cfg_discovery to ensure that
Cargo handles this properly. Unlike the other tests, there is no
expected output and an exit code 0.